统计模块初步测试后的更新

+++ agent/src/perf/scalar/CpuUsageScalar.h
@@ -15,7 +15,7 @@ public:
 
 
     virtual void UpdateThrod();
-    virtual int FetchData();
+    virtual double FetchData();
 
 
 private:
