+++ agent/src/perf/scalar/MainFSDiskUsageScalar.h
@@ -18,7 +18,7 @@ public:
 
 
     virtual void UpdateThrod();
-    virtual int FetchData();
+    virtual double FetchData();
 
 
 private:
