+++ agent/src/perf/scalar/SwapUsageScalar.h
@@ -17,7 +17,7 @@ public:
 
 
     virtual void UpdateThrod();
-    virtual int FetchData();
+    virtual double FetchData();
 
 
 private:
